+++ packages/github_auth/lib/github_auth.dart
@@ -1,7 +1,11 @@
 library github_auth;
 
 export 'package:flutter_auth_core/flutter_auth_core.dart'
-    show FlutterAuthResult, FlutterAuthException, FlutterAuthExceptionCode;
+    show
+        FlutterAuth,
+        FlutterAuthResult,
+        FlutterAuthException,
+        FlutterAuthExceptionCode;
 
 export 'src/github_auth.dart';
 export 'src/github_api_error.dart';