+++ packages/twitter_auth/lib/twitter_auth.dart
@@ -4,7 +4,11 @@ library twitter_auth;
 //     show TwitterAuthPlatform, TwitterAuthResult;
 
 export 'package:flutter_auth_core/flutter_auth_core.dart'
-    show FlutterAuthResult, FlutterAuthException, FlutterAuthExceptionCode;
+    show
+        FlutterAuth,
+        FlutterAuthResult,
+        FlutterAuthException,
+        FlutterAuthExceptionCode;
 
 export 'src/twitter_auth.dart';
 export 'src/twitter_api_error.dart';