+++ packages/flutter_auth_core/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.3
+
+- **FEAT**: export FlutterAuth class
+
+## 0.0.2
+
+- **DOCS**: update meta
+
 ## 0.0.1
 
 * Initial Open Source release.